projects/fintech_project/requirements.txt
certifi==2022.12.7
charset-normalizer==2.1.1
Django==2.2
djangorestframework==3.9.1
idna==3.4
Naked==0.1.32
pycryptodome==3.16.0
pycryptodomex==3.16.0
pytz==2022.6
PyYAML==6.0
requests==2.28.1
shellescape==3.8.1
sqlparse==0.4.3
urllib3==1.26.13
